storage - fix layers to be able to use native storage service in sandbox
