+++ src/vs/code/electron-browser/sharedProcess/sharedProcessMain.ts
@@ -58,8 +58,8 @@ import { INativeHostService } from 'vs/platform/native/electron-sandbox/native';
 import { LoggerService } from 'vs/platform/log/node/loggerService';
 import { UserDataSyncLogService } from 'vs/platform/userDataSync/common/userDataSyncLog';
 import { UserDataAutoSyncService } from 'vs/platform/userDataSync/electron-sandbox/userDataAutoSyncService';
-import { NativeStorageService2 } from 'vs/platform/storage/node/storageService2';
-import { StorageDatabaseChannelClient } from 'vs/platform/storage/node/storageIpc';
+import { NativeStorageService2 } from 'vs/platform/storage/electron-sandbox/storageService2';
+import { StorageDatabaseChannelClient } from 'vs/platform/storage/common/storageIpc';
 import { IStorageService } from 'vs/platform/storage/common/storage';
 import { GlobalExtensionEnablementService } from 'vs/platform/extensionManagement/common/extensionEnablementService';
 import { UserDataSyncResourceEnablementService } from 'vs/platform/userDataSync/common/userDataSyncResourceEnablementService';
