+++ src/vs/code/electron-main/app.ts
@@ -58,8 +58,8 @@ import { joinPath } from 'vs/base/common/resources';
 import { localize } from 'vs/nls';
 import { Schemas } from 'vs/base/common/network';
 import { SnapUpdateService } from 'vs/platform/update/electron-main/updateService.snap';
-import { IStorageMainService, StorageMainService } from 'vs/platform/storage/node/storageMainService';
-import { StorageDatabaseChannel } from 'vs/platform/storage/node/storageIpc';
+import { IStorageMainService, StorageMainService } from 'vs/platform/storage/electron-main/storageMainService';
+import { StorageDatabaseChannel } from 'vs/platform/storage/electron-main/storageIpc';
 import { BackupMainService } from 'vs/platform/backup/electron-main/backupMainService';
 import { IBackupMainService } from 'vs/platform/backup/electron-main/backup';
 import { WorkspacesHistoryMainService, IWorkspacesHistoryMainService } from 'vs/platform/workspaces/electron-main/workspacesHistoryMainService';
