+++ src/vs/platform/storage/electron-main/storageMain.ts
