+++ src/vs/platform/storage/electron-sandbox/storageService2.ts
