+++ src/vs/workbench/electron-sandbox/sandbox.simpleservices.ts
@@ -6,7 +6,6 @@
 /* eslint-disable code-no-standalone-editor */
 /* eslint-disable code-import-patterns */
 
-import { ConsoleLogger, LogService } from 'vs/platform/log/common/log';
 import { ISignService } from 'vs/platform/sign/common/sign';
 import { URI } from 'vs/base/common/uri';
 import { InMemoryFileSystemProvider } from 'vs/platform/files/common/inMemoryFilesystemProvider';
@@ -15,7 +14,6 @@ import { IAddressProvider } from 'vs/platform/remote/common/remoteAgentConnectio
 import { ITelemetryData, ITelemetryInfo, ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IExtension } from 'vs/platform/extensions/common/extensions';
 import { SimpleConfigurationService as BaseSimpleConfigurationService } from 'vs/editor/standalone/browser/simpleServices';
-import { InMemoryStorageService } from 'vs/platform/storage/common/storage';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IBackupFileService, IResolvedBackup } from 'vs/workbench/services/backup/common/backup';
 import { ITextSnapshot } from 'vs/editor/common/model';
@@ -179,13 +177,6 @@ export class SimpleWorkspaceService implements IWorkspaceContextService {
 
 //#region Configuration
 
-export class SimpleStorageService extends InMemoryStorageService { }
-
-//#endregion
-
-
-//#region Configuration
-
 export class SimpleConfigurationService extends BaseSimpleConfigurationService implements IWorkbenchConfigurationService {
 	async whenRemoteConfigurationLoaded() { }
 }
@@ -193,15 +184,7 @@ export class SimpleConfigurationService extends BaseSimpleConfigurationService i
 //#endregion
 
 
-//#region Logger
-
-export class SimpleLogService extends LogService {
-
-	constructor() {
-		super(new ConsoleLogger());
-	}
-
-}
+//#region Signing
 
 export class SimpleSignService implements ISignService {
